+++ .travis.sh
@@ -2,6 +2,9 @@
 
 set -ex
 
+./build_openfst.sh
+./run_openfst.sh
+
 cargo build --all
 cargo test --all
 cargo check --benches --all # running benches on travis is useless