+++ rustfst/src/fst_traits/mod.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 mod macros;
+mod binary_parser;
 mod expanded_fst;
 mod final_states_iterator;
 mod fst;
@@ -7,6 +8,7 @@ mod mutable_fst;
 mod paths_iterator;
 mod text_parser;
 
+pub use self::binary_parser::BinaryParser;
 pub use self::expanded_fst::ExpandedFst;
 pub use self::final_states_iterator::FinalStatesIterator;
 pub use self::fst::{ArcIterator, CoreFst, Fst, StateIterator};