+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `find_strongly_connected_components`  which is the implementation of the Tarjan's algoritm to find the strongly connect components in a directed graph.
 - Add the `FstProperties` bitflags struct and a function to commpute the property flags for an `Fst`.
 - Implement `topsort` and `statesort`.
+- Add `BinaryParser` trait. Should be used to parse an Fst in binary format.
+- Implement `BinaryParser` for VectorFst i.e VectorFst binary deserialization is now supported.
 
 ### Changed
 - Before test cases were generated with pynini (python wrapper around openfst). Now they are directly generated with OpenFST (c++). Allows to test operations that are not wrapped.
+- Test cases are now generated directly in the CI by buiding and running openfst which allows to avoid pushing data on the repo.
 
 ## [0.3.0] - 2019-04-03